Modified go-target()
implement fuzzy logic alternative method in go-target(), works when light is far from the robot.
If the light is too close, fails to detect the light. 
One thing to improves: in rotate_while_scan modify the condition to make the robot to turn to face the center (use difference between RR and LL ).

In the old method, modified slightly_off condition using abs(RRAve-LLAve), works well when testing individual behavior, and when no obstacle met.
Phototransistors read funny values (1023) sometimes, which mess up the slightly_off behaviour. (which is the reason why I give up on the accuracy and uses fuzzy logic )